FIles separated for render & settings
