+++ classes/class-uagb-init-blocks.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * UAGB Blocks Initializer
  *
@@ -8,7 +9,7 @@
  * @package UAGB
  */
 
-if ( ! defined( 'ABSPATH' ) ) {
+if (!defined('ABSPATH')) {
 	exit; // Exit if accessed directly.
 }
 
@@ -17,7 +18,8 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @package UAGB
  */
-class UAGB_Init_Blocks {
+class UAGB_Init_Blocks
+{
 
 
 	/**
@@ -30,8 +32,9 @@ class UAGB_Init_Blocks {
 	/**
 	 *  Initiator
 	 */
-	public static function get_instance() {
-		if ( ! isset( self::$instance ) ) {
+	public static function get_instance()
+	{
+		if (!isset(self::$instance)) {
 			self::$instance = new self();
 		}
 		return self::$instance;
@@ -40,23 +43,24 @@ class UAGB_Init_Blocks {
 	/**
 	 * Constructor
 	 */
-	public function __construct() {
+	public function __construct()
+	{
 		// Hook: Frontend assets.
-		add_action( 'enqueue_block_assets', array( $this, 'block_assets' ) );
+		add_action('enqueue_block_assets', array($this, 'block_assets'));
 
 		// Hook: Editor assets.
-		add_action( 'enqueue_block_editor_assets', array( $this, 'editor_assets' ) );
+		add_action('enqueue_block_editor_assets', array($this, 'editor_assets'));
 
-		add_filter( 'block_categories', array( $this, 'register_block_category' ), 10, 2 );
+		add_filter('block_categories', array($this, 'register_block_category'), 10, 2);
 
-		add_action( 'wp_ajax_uagb_gf_shortcode', array( $this, 'gf_shortcode' ) );
-		add_action( 'wp_ajax_nopriv_uagb_gf_shortcode', array( $this, 'gf_shortcode' ) );
+		add_action('wp_ajax_uagb_gf_shortcode', array($this, 'gf_shortcode'));
+		add_action('wp_ajax_nopriv_uagb_gf_shortcode', array($this, 'gf_shortcode'));
 
-		add_action( 'wp_ajax_uagb_cf7_shortcode', array( $this, 'cf7_shortcode' ) );
-		add_action( 'wp_ajax_nopriv_uagb_cf7_shortcode', array( $this, 'cf7_shortcode' ) );
+		add_action('wp_ajax_uagb_cf7_shortcode', array($this, 'cf7_shortcode'));
+		add_action('wp_ajax_nopriv_uagb_cf7_shortcode', array($this, 'cf7_shortcode'));
 
-		if ( ! is_admin() ) {
-			add_action( 'render_block', array( $this, 'render_block' ), 5, 2 );
+		if (!is_admin()) {
+			add_action('render_block', array($this, 'render_block'), 5, 2);
 		}
 	}
 	/**
@@ -67,28 +71,29 @@ class UAGB_Init_Blocks {
 	 * @since 1.21.0
 	 * @return mixed Returns the new block content.
 	 */
-	public function render_block( $block_content, $block ) {
+	public function render_block($block_content, $block)
+	{
 
 		$block_attributes = $block['attrs'];
 
-		if ( isset( $block_attributes['UAGDisplayConditions'] ) && array_key_exists( 'UAGDisplayConditions', $block_attributes ) ) {
+		if (isset($block_attributes['UAGDisplayConditions']) && array_key_exists('UAGDisplayConditions', $block_attributes)) {
 
-			switch ( $block_attributes['UAGDisplayConditions'] ) {
+			switch ($block_attributes['UAGDisplayConditions']) {
 
 				case 'userstate':
-					$block_content = $this->user_state_visibility( $block_attributes, $block_content );
+					$block_content = $this->user_state_visibility($block_attributes, $block_content);
 					break;
 
 				case 'userRole':
-					$block_content = $this->user_role_visibility( $block_attributes, $block_content );
+					$block_content = $this->user_role_visibility($block_attributes, $block_content);
 					break;
 
 				case 'browser':
-					$block_content = $this->browser_visibility( $block_attributes, $block_content );
+					$block_content = $this->browser_visibility($block_attributes, $block_content);
 					break;
 
 				case 'os':
-					$block_content = $this->os_visibility( $block_attributes, $block_content );
+					$block_content = $this->os_visibility($block_attributes, $block_content);
 					break;
 
 				default:
@@ -107,15 +112,16 @@ class UAGB_Init_Blocks {
 	 * @since 1.21.0
 	 * @return mixed Returns the new block content.
 	 */
-	public function user_role_visibility( $block_attributes, $block_content ) {
+	public function user_role_visibility($block_attributes, $block_content)
+	{
 
 		$user = wp_get_current_user();
 
-		if ( isset( $block_attributes['UAGUserRole'] ) && array_key_exists( 'UAGUserRole', $block_attributes ) ) {
+		if (isset($block_attributes['UAGUserRole']) && array_key_exists('UAGUserRole', $block_attributes)) {
 
 			$value = $block_attributes['UAGUserRole'];
 
-			if ( is_user_logged_in() && in_array( $value, $user->roles, true ) ) {
+			if (is_user_logged_in() && in_array($value, $user->roles, true)) {
 				return '';
 			}
 		}
@@ -129,9 +135,10 @@ class UAGB_Init_Blocks {
 	 * @since 1.21.0
 	 * @return mixed Returns the new block content.
 	 */
-	public function os_visibility( $block_attributes, $block_content ) {
+	public function os_visibility($block_attributes, $block_content)
+	{
 
-		if ( ! array_key_exists( 'UAGSystem', $block_attributes ) ) {
+		if (!array_key_exists('UAGSystem', $block_attributes)) {
 			return $block_content;
 		}
 
@@ -147,7 +154,7 @@ class UAGB_Init_Blocks {
 			'mac_os'   => '(Mac_PowerPC)|(Macintosh)',
 		);
 
-		if ( preg_match( '@' . $os[ $value ] . '@', $_SERVER['HTTP_USER_AGENT'] ) ) {
+		if (preg_match('@' . $os[$value] . '@', $_SERVER['HTTP_USER_AGENT'])) {
 			return '';
 		}
 
@@ -162,9 +169,10 @@ class UAGB_Init_Blocks {
 	 * @since 1.21.0
 	 * @return mixed Returns the new block content.
 	 */
-	public function browser_visibility( $block_attributes, $block_content ) {
+	public function browser_visibility($block_attributes, $block_content)
+	{
 
-		if ( ! array_key_exists( 'UAGBrowser', $block_attributes ) ) {
+		if (!array_key_exists('UAGBrowser', $block_attributes)) {
 			return $block_content;
 		}
 
@@ -184,24 +192,24 @@ class UAGB_Init_Blocks {
 
 		$show = false;
 
-		if ( 'ie' === $value ) {
-			if ( false !== strpos( $_SERVER['HTTP_USER_AGENT'], $browsers[ $value ][0] ) || false !== strpos( $_SERVER['HTTP_USER_AGENT'], $browsers[ $value ][1] ) ) {
+		if ('ie' === $value) {
+			if (false !== strpos($_SERVER['HTTP_USER_AGENT'], $browsers[$value][0]) || false !== strpos($_SERVER['HTTP_USER_AGENT'], $browsers[$value][1])) {
 				$show = true;
 			}
 		} else {
-			if ( false !== strpos( $_SERVER['HTTP_USER_AGENT'], $browsers[ $value ] ) ) {
+			if (false !== strpos($_SERVER['HTTP_USER_AGENT'], $browsers[$value])) {
 				$show = true;
 
 				// Additional check for Chrome that returns Safari.
-				if ( 'safari' === $value || 'firefox' === $value ) {
-					if ( false !== strpos( $_SERVER['HTTP_USER_AGENT'], 'Chrome' ) ) {
+				if ('safari' === $value || 'firefox' === $value) {
+					if (false !== strpos($_SERVER['HTTP_USER_AGENT'], 'Chrome')) {
 						$show = false;
 					}
 				}
 			}
 		}
 
-		return ( $show ) ? '' : $block_content;
+		return ($show) ? '' : $block_content;
 	}
 	/**
 	 * User State Visibility.
@@ -212,18 +220,18 @@ class UAGB_Init_Blocks {
 	 * @since 1.21.0
 	 * @return mixed Returns the new block content.
 	 */
-	public function user_state_visibility( $block_attributes, $block_content ) {
+	public function user_state_visibility($block_attributes, $block_content)
+	{
 
-		if ( isset( $block_attributes['UAGLoggedIn'] ) && $block_attributes['UAGLoggedIn'] && is_user_logged_in() ) {
+		if (isset($block_attributes['UAGLoggedIn']) && $block_attributes['UAGLoggedIn'] && is_user_logged_in()) {
 			return '';
 		}
 
-		if ( isset( $block_attributes['UAGLoggedOut'] ) && $block_attributes['UAGLoggedOut'] && ! is_user_logged_in() ) {
+		if (isset($block_attributes['UAGLoggedOut']) && $block_attributes['UAGLoggedOut'] && !is_user_logged_in()) {
 			return '';
 		}
 
 		return $block_content;
-
 	}
 
 	/**
@@ -231,18 +239,19 @@ class UAGB_Init_Blocks {
 	 *
 	 * @since 1.12.0
 	 */
-	public function gf_shortcode() {
+	public function gf_shortcode()
+	{
 
-		check_ajax_referer( 'uagb_ajax_nonce', 'nonce' );
+		check_ajax_referer('uagb_ajax_nonce', 'nonce');
 
-		$id = intval( $_POST['formId'] );
+		$id = intval($_POST['formId']);
 
-		if ( $id && 0 !== $id && -1 !== $id ) {
-			$data['html'] = do_shortcode( '[gravityforms id="' . $id . '" ajax="true"]' );
+		if ($id && 0 !== $id && -1 !== $id) {
+			$data['html'] = do_shortcode('[gravityforms id="' . $id . '" ajax="true"]');
 		} else {
-			$data['html'] = '<p>' . __( 'Please select a valid Gravity Form.', 'ultimate-addons-for-gutenberg' ) . '</p>';
+			$data['html'] = '<p>' . __('Please select a valid Gravity Form.', 'ultimate-addons-for-gutenberg') . '</p>';
 		}
-		wp_send_json_success( $data );
+		wp_send_json_success($data);
 	}
 
 	/**
@@ -250,18 +259,19 @@ class UAGB_Init_Blocks {
 	 *
 	 * @since 1.10.0
 	 */
-	public function cf7_shortcode() {
+	public function cf7_shortcode()
+	{
 
-		check_ajax_referer( 'uagb_ajax_nonce', 'nonce' );
+		check_ajax_referer('uagb_ajax_nonce', 'nonce');
 
-		$id = intval( $_POST['formId'] );
+		$id = intval($_POST['formId']);
 
-		if ( $id && 0 !== $id && -1 !== $id ) {
-			$data['html'] = do_shortcode( '[contact-form-7 id="' . $id . '" ajax="true"]' );
+		if ($id && 0 !== $id && -1 !== $id) {
+			$data['html'] = do_shortcode('[contact-form-7 id="' . $id . '" ajax="true"]');
 		} else {
-			$data['html'] = '<p>' . __( 'Please select a valid Contact Form 7.', 'ultimate-addons-for-gutenberg' ) . '</p>';
+			$data['html'] = '<p>' . __('Please select a valid Contact Form 7.', 'ultimate-addons-for-gutenberg') . '</p>';
 		}
-		wp_send_json_success( $data );
+		wp_send_json_success($data);
 	}
 
 	/**
@@ -271,13 +281,14 @@ class UAGB_Init_Blocks {
 	 * @param object $post Post object.
 	 * @since 1.0.0
 	 */
-	public function register_block_category( $categories, $post ) {
+	public function register_block_category($categories, $post)
+	{
 		return array_merge(
 			$categories,
 			array(
 				array(
 					'slug'  => 'uagb',
-					'title' => __( 'Ultimate Addons Blocks', 'ultimate-addons-for-gutenberg' ),
+					'title' => __('Ultimate Addons Blocks', 'ultimate-addons-for-gutenberg'),
 				),
 			)
 		);
@@ -288,13 +299,14 @@ class UAGB_Init_Blocks {
 	 *
 	 * @since 1.0.0
 	 */
-	public function block_assets() {
+	public function block_assets()
+	{
 
-		if ( ! is_admin() ) {
+		if (!is_admin()) {
 
-			if ( class_exists( 'WooCommerce' ) ) {
+			if (class_exists('WooCommerce')) {
 
-				if ( false === UAGB_Helper::$uag_flag ) {
+				if (false === UAGB_Helper::$uag_flag) {
 					return;
 				}
 			} else {
@@ -306,13 +318,13 @@ class UAGB_Init_Blocks {
 				 *
 				 * @param \WP_Post $post The global post.
 				 */
-				$post = apply_filters( 'uagb_post_for_stylesheet', $post );
+				$post = apply_filters('uagb_post_for_stylesheet', $post);
 
-				if ( false === has_blocks( $post ) ) {
+				if (false === has_blocks($post)) {
 					return;
 				}
 
-				if ( false === UAGB_Helper::$uag_flag ) {
+				if (false === UAGB_Helper::$uag_flag) {
 					return;
 				}
 			}
@@ -325,7 +337,7 @@ class UAGB_Init_Blocks {
 			UAGB_VER
 		);
 
-		if ( is_rtl() ) {
+		if (is_rtl()) {
 			wp_enqueue_style(
 				'uagb-style-rtl', // Handle.
 				UAGB_URL . 'dist/blocks.style.rtl.css', // RTL style CSS.
@@ -335,81 +347,80 @@ class UAGB_Init_Blocks {
 		}
 
 		$blocks          = UAGB_Config::get_block_attributes();
-		$disabled_blocks = UAGB_Admin_Helper::get_admin_settings_option( '_uagb_blocks', array() );
+		$disabled_blocks = UAGB_Admin_Helper::get_admin_settings_option('_uagb_blocks', array());
 		$block_assets    = UAGB_Config::get_block_assets();
 
-		foreach ( $blocks as $slug => $value ) {
-			$_slug = str_replace( 'uagb/', '', $slug );
+		foreach ($blocks as $slug => $value) {
+			$_slug = str_replace('uagb/', '', $slug);
 
-			if ( ! ( isset( $disabled_blocks[ $_slug ] ) && 'disabled' === $disabled_blocks[ $_slug ] ) ) {
+			if (!(isset($disabled_blocks[$_slug]) && 'disabled' === $disabled_blocks[$_slug])) {
 
-				$js_assets = ( isset( $blocks[ $slug ]['js_assets'] ) ) ? $blocks[ $slug ]['js_assets'] : array();
+				$js_assets = (isset($blocks[$slug]['js_assets'])) ? $blocks[$slug]['js_assets'] : array();
 
-				$css_assets = ( isset( $blocks[ $slug ]['css_assets'] ) ) ? $blocks[ $slug ]['css_assets'] : array();
+				$css_assets = (isset($blocks[$slug]['css_assets'])) ? $blocks[$slug]['css_assets'] : array();
 
-				if ( 'cf7-styler' === $_slug ) {
-					if ( ! wp_script_is( 'contact-form-7', 'enqueued' ) ) {
-						wp_enqueue_script( 'contact-form-7' );
+				if ('cf7-styler' === $_slug) {
+					if (!wp_script_is('contact-form-7', 'enqueued')) {
+						wp_enqueue_script('contact-form-7');
 					}
 
-					if ( ! wp_script_is( ' wpcf7-admin', 'enqueued' ) ) {
-						wp_enqueue_script( ' wpcf7-admin' );
+					if (!wp_script_is(' wpcf7-admin', 'enqueued')) {
+						wp_enqueue_script(' wpcf7-admin');
 					}
 				}
 
-				foreach ( $js_assets as $asset_handle => $val ) {
+				foreach ($js_assets as $asset_handle => $val) {
 					// Scripts.
 					wp_register_script(
 						$val, // Handle.
-						$block_assets[ $val ]['src'],
-						$block_assets[ $val ]['dep'],
+						$block_assets[$val]['src'],
+						$block_assets[$val]['dep'],
 						UAGB_VER,
 						true
 					);
 
-					$skip_editor = isset( $block_assets[ $val ]['skipEditor'] ) ? $block_assets[ $val ]['skipEditor'] : false;
+					$skip_editor = isset($block_assets[$val]['skipEditor']) ? $block_assets[$val]['skipEditor'] : false;
 
-					if ( is_admin() && false === $skip_editor ) {
-						wp_enqueue_script( $val );
+					if (is_admin() && false === $skip_editor) {
+						wp_enqueue_script($val);
 					}
 				}
 
-				foreach ( $css_assets as $asset_handle => $val ) {
+				foreach ($css_assets as $asset_handle => $val) {
 					// Styles.
 					wp_register_style(
 						$val, // Handle.
-						$block_assets[ $val ]['src'],
-						$block_assets[ $val ]['dep'],
+						$block_assets[$val]['src'],
+						$block_assets[$val]['dep'],
 						UAGB_VER
 					);
 
-					if ( is_admin() ) {
-						wp_enqueue_style( $val );
+					if (is_admin()) {
+						wp_enqueue_style($val);
 					}
 				}
 			}
 		}
 
-		$uagb_masonry_ajax_nonce = wp_create_nonce( 'uagb_masonry_ajax_nonce' );
+		$uagb_masonry_ajax_nonce = wp_create_nonce('uagb_masonry_ajax_nonce');
 		wp_localize_script(
 			'uagb-post-js',
 			'uagb_data',
 			array(
-				'ajax_url'                => admin_url( 'admin-ajax.php' ),
+				'ajax_url'                => admin_url('admin-ajax.php'),
 				'uagb_masonry_ajax_nonce' => $uagb_masonry_ajax_nonce,
 			)
 		);
 
-		$uagb_forms_ajax_nonce = wp_create_nonce( 'uagb_forms_ajax_nonce' );
+		$uagb_forms_ajax_nonce = wp_create_nonce('uagb_forms_ajax_nonce');
 		wp_localize_script(
 			'uagb-forms-js',
 			'uagb_forms_data',
 			array(
-				'ajax_url'              => admin_url( 'admin-ajax.php' ),
+				'ajax_url'              => admin_url('admin-ajax.php'),
 				'uagb_forms_ajax_nonce' => $uagb_forms_ajax_nonce,
 			)
 		);
-
 	} // End function editor_assets().
 
 	/**
@@ -417,25 +428,26 @@ class UAGB_Init_Blocks {
 	 *
 	 * @since 1.0.0
 	 */
-	public function editor_assets() {
+	public function editor_assets()
+	{
 
-		$uagb_ajax_nonce = wp_create_nonce( 'uagb_ajax_nonce' );
+		$uagb_ajax_nonce = wp_create_nonce('uagb_ajax_nonce');
 		// Scripts.
 		wp_enqueue_script(
 			'uagb-block-editor-js', // Handle.
 			UAGB_URL . 'dist/build/blocks.js',
-			array( 'wp-blocks', 'wp-i18n', 'wp-element', 'wp-components', 'wp-editor', 'wp-api-fetch' ), // Dependencies, defined above.
+			array('wp-blocks', 'wp-i18n', 'wp-element', 'wp-components', 'wp-editor', 'wp-api-fetch'), // Dependencies, defined above.
 			UAGB_VER,
 			true // Enqueue the script in the footer.
 		);
 
-		wp_set_script_translations( 'uagb-block-editor-js', 'ultimate-addons-for-gutenberg' );
+		wp_set_script_translations('uagb-block-editor-js', 'ultimate-addons-for-gutenberg');
 
 		// Styles.
 		wp_enqueue_style(
 			'uagb-block-editor-css', // Handle.
 			UAGB_URL . 'dist/build/blocks.css', // Block editor CSS.
-			array( 'wp-edit-blocks' ), // Dependency to include the CSS after it.
+			array('wp-edit-blocks'), // Dependency to include the CSS after it.
 			UAGB_VER
 		);
 
@@ -443,31 +455,31 @@ class UAGB_Init_Blocks {
 		wp_enqueue_style(
 			'uagb-block-common-editor-css', // Handle.
 			UAGB_URL . 'dist/blocks.commoneditorstyle.build.css', // Block editor CSS.
-			array( 'wp-edit-blocks' ), // Dependency to include the CSS after it.
+			array('wp-edit-blocks'), // Dependency to include the CSS after it.
 			UAGB_VER
 		);
 
-		wp_enqueue_script( 'uagb-deactivate-block-js', UAGB_URL . 'dist/blocks-deactivate.js', array( 'wp-blocks' ), UAGB_VER, true );
+		wp_enqueue_script('uagb-deactivate-block-js', UAGB_URL . 'dist/blocks-deactivate.js', array('wp-blocks'), UAGB_VER, true);
 
 		$blocks       = array();
-		$saved_blocks = UAGB_Admin_Helper::get_admin_settings_option( '_uagb_blocks' );
+		$saved_blocks = UAGB_Admin_Helper::get_admin_settings_option('_uagb_blocks');
 
-		if ( is_array( $saved_blocks ) ) {
-			foreach ( $saved_blocks as $slug => $data ) {
+		if (is_array($saved_blocks)) {
+			foreach ($saved_blocks as $slug => $data) {
 				$_slug         = 'uagb/' . $slug;
-				$current_block = UAGB_Config::$block_attributes[ $_slug ];
+				$current_block = UAGB_Config::$block_attributes[$_slug];
 
-				if ( isset( $current_block['is_child'] ) && $current_block['is_child'] ) {
+				if (isset($current_block['is_child']) && $current_block['is_child']) {
 					continue;
 				}
 
-				if ( isset( $current_block['is_active'] ) && ! $current_block['is_active'] ) {
+				if (isset($current_block['is_active']) && !$current_block['is_active']) {
 					continue;
 				}
 
-				if ( isset( $saved_blocks[ $slug ] ) ) {
-					if ( 'disabled' === $saved_blocks[ $slug ] ) {
-						array_push( $blocks, $_slug );
+				if (isset($saved_blocks[$slug])) {
+					if ('disabled' === $saved_blocks[$slug]) {
+						array_push($blocks, $_slug);
 					}
 				}
 			}
@@ -487,7 +499,7 @@ class UAGB_Init_Blocks {
 			array(
 				'blocks'                 => UAGB_Config::get_block_attributes(),
 				'category'               => 'uagb',
-				'ajax_url'               => admin_url( 'admin-ajax.php' ),
+				'ajax_url'               => admin_url('admin-ajax.php'),
 				'cf7_forms'              => $this->get_cf7_forms(),
 				'gf_forms'               => $this->get_gravity_forms(),
 				'tablet_breakpoint'      => UAGB_TABLET_BREAKPOINT,
@@ -502,7 +514,7 @@ class UAGB_Init_Blocks {
 				'uagb_url'               => UAGB_URL,
 				'uagb_mime_type'         => UAGB_Helper::get_mime_type(),
 				'uagb_site_url'          => UAGB_URI,
-				'uagb_display_condition' => apply_filters( 'enable_block_condition', true ),
+				'uagb_display_condition' => apply_filters('enable_block_condition', true),
 			)
 		);
 	} // End function editor_assets().
@@ -511,7 +523,8 @@ class UAGB_Init_Blocks {
 	 *
 	 *  @since 1.21.0
 	 */
-	public function get_user_role() {
+	public function get_user_role()
+	{
 
 		global $wp_roles;
 
@@ -521,10 +534,10 @@ class UAGB_Init_Blocks {
 
 		$field_options[0] = array(
 			'value' => '',
-			'label' => __( 'None', 'ultimate-addons-for-gutenberg' ),
+			'label' => __('None', 'ultimate-addons-for-gutenberg'),
 		);
 
-		foreach ( $role_lists as $key => $role_list ) {
+		foreach ($role_lists as $key => $role_list) {
 			$field_options[] = array(
 				'value' => $key,
 				'label' => $role_list,
@@ -539,21 +552,22 @@ class UAGB_Init_Blocks {
 	 *
 	 * @since 1.10.0
 	 */
-	public function get_cf7_forms() {
+	public function get_cf7_forms()
+	{
 		$field_options = array();
 
-		if ( class_exists( 'WPCF7_ContactForm' ) ) {
+		if (class_exists('WPCF7_ContactForm')) {
 			$args             = array(
 				'post_type'      => 'wpcf7_contact_form',
 				'posts_per_page' => -1,
 			);
-			$forms            = get_posts( $args );
+			$forms            = get_posts($args);
 			$field_options[0] = array(
 				'value' => -1,
-				'label' => __( 'Select Form', 'ultimate-addons-for-gutenberg' ),
+				'label' => __('Select Form', 'ultimate-addons-for-gutenberg'),
 			);
-			if ( $forms ) {
-				foreach ( $forms as $form ) {
+			if ($forms) {
+				foreach ($forms as $form) {
 					$field_options[] = array(
 						'value' => $form->ID,
 						'label' => $form->post_title,
@@ -562,9 +576,9 @@ class UAGB_Init_Blocks {
 			}
 		}
 
-		if ( empty( $field_options ) ) {
+		if (empty($field_options)) {
 			$field_options = array(
-				'-1' => __( 'You have not added any Contact Form 7 yet.', 'ultimate-addons-for-gutenberg' ),
+				'-1' => __('You have not added any Contact Form 7 yet.', 'ultimate-addons-for-gutenberg'),
 			);
 		}
 		return $field_options;
@@ -576,17 +590,18 @@ class UAGB_Init_Blocks {
 	 * @since 1.12.0
 	 * @return array Key Value paired array.
 	 */
-	public function get_gravity_forms() {
+	public function get_gravity_forms()
+	{
 		$field_options = array();
 
-		if ( class_exists( 'GFForms' ) ) {
-			$forms            = RGFormsModel::get_forms( null, 'title' );
+		if (class_exists('GFForms')) {
+			$forms            = RGFormsModel::get_forms(null, 'title');
 			$field_options[0] = array(
 				'value' => -1,
-				'label' => __( 'Select Form', 'ultimate-addons-for-gutenberg' ),
+				'label' => __('Select Form', 'ultimate-addons-for-gutenberg'),
 			);
-			if ( is_array( $forms ) ) {
-				foreach ( $forms as $form ) {
+			if (is_array($forms)) {
+				foreach ($forms as $form) {
 					$field_options[] = array(
 						'value' => $form->id,
 						'label' => $form->title,
@@ -595,9 +610,9 @@ class UAGB_Init_Blocks {
 			}
 		}
 
-		if ( empty( $field_options ) ) {
+		if (empty($field_options)) {
 			$field_options = array(
-				'-1' => __( 'You have not added any Gravity Forms yet.', 'ultimate-addons-for-gutenberg' ),
+				'-1' => __('You have not added any Gravity Forms yet.', 'ultimate-addons-for-gutenberg'),
 			);
 		}
