+++ package.json
@@ -63,6 +63,8 @@
     "check-engines": "wp-scripts check-engines",
     "makepot": "vendor/bin/wp i18n make-pot . --skip-audit --exclude='.dev,.github,.wordpress-org,build,docs,dist,node_modules,vendor,wordpress' languages/ultimate-addons-for-gutenberg.pot",
     "package": "rm *.zip && wp dist-archive . ultimate-addons-for-gutenberg",
-    "packages-update": "wp-scripts packages-update"
+    "packages-update": "wp-scripts packages-update",
+    "lint:js:columns": "wp-scripts lint-js ./src/blocks/columns",
+    "lint:js:fix:columns": "wp-scripts lint-js ./src/blocks/columns --fix"
   }
 }
